ttsd-user-rest-service/requirements.txt
click==6.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
redis==2.10.5
SQLAlchemy==1.0.14
Werkzeug==0.11.10
WTForms==2.1
https://github.com/ttsdzzg/python-repository/raw/master/mns-1.1.4.zip#egg=mns==1.1.4
python-etcd==0.4.5
